+++ pkg/config/adapter_test.go
