+++ pkg/config/doc.go
