+++ pkg/errors/doc.go
