+++ pkg/git/test_helpers.go
@@ -2,7 +2,7 @@ package git
 
 import (
 	"context"
-	"github.com/bashhack/gitbak/internal/logger"
+	"github.com/bashhack/gitbak/pkg/logger"
 	"os"
 	"os/exec"
 	"path/filepath"
