+++ pkg/lock/doc.go
